Move renderers under simplates
