+++ aspen/renderers.py
@@ -1,0 +1,5 @@
+
+# for backwards compatibility with aspen-renderer modules
+from .simplates.renderers import Factory, Renderer
+
+Factory, Renderer # make pyflakes happy
