+++ aspen/simplates/renderers/json_dump.py
@@ -1,6 +1,6 @@
 """
-aspen.renderers.json_dump
-~~~~~~~~~~~~~~~~~~~~~~~~~
+aspen.simplates.renderers.json_dump
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 """
 from __future__ import absolute_import
 from __future__ import division
@@ -8,7 +8,7 @@ from __future__ import print_function
 from __future__ import unicode_literals
 
 from . import Renderer, Factory
-from .. import json
+from ... import json
 
 class Renderer(Renderer):
     def compile(self, filepath, raw):
